+++ src/constants.ts
@@ -1,6 +1,7 @@
 // Action events
 export const EVENT_PULL_REQUEST = "pull_request";
 export const EVENT_PUSH = "push";
+export const EVENT_SCHEDULE = "schedule";
 
 // Action configuration
 export const INPUT_NAME_COMMENT = "comment";
@@ -44,7 +45,11 @@ export const UTF8 = "utf-8";
 export const COMMIT_MODE_FILE = "100644";
 export const COMMIT_TYPE_BLOB = "blob";
 
+export const GIT_OBJECT_TYPE_DIR = "dir";
+export const GIT_OBJECT_TYPE_FILE = "file";
+
 export const STATUS_ADDED = "added";
+export const STATUS_EXISTS = "exists";
 export const STATUS_MODIFIED = "modified";
 export const STATUS_REMOVED = "removed";