+++ src/github-api.ts
@@ -1,47 +1,50 @@
 import * as github from "@actions/github";
 import { Octokit } from "@octokit/core";
-import { GitGetCommitResponseData } from "@octokit/types";
+import {
+  GitGetCommitResponseData,
+  ReposGetContentResponseData,
+} from "@octokit/types";
 
 import { COMMIT_MODE_FILE, COMMIT_TYPE_BLOB, PR_NOT_FOUND } from "./constants";
-import { CommitInfo, GitBlob, GitFileData, GitFileInfo } from "./types";
+import {
+  CommitInfo,
+  GitBlob,
+  GitFileData,
+  GitFileInfo,
+  GitObjectInfo,
+} from "./types";
 
+
+type DirContents = ReposGetContentResponseData[];
+
+type FileContents = ReposGetContentResponseData;
 
 type GitCommit = GitGetCommitResponseData;
 
+type RepoContents = FileContents | DirContents;
+
+
+const owner = github.context.repo.owner;
+const repo = github.context.repo.repo;
 
 async function getCommitAt(client: Octokit, ref: string): Promise<GitCommit> {
-  const { data: refData } = await client.git.getRef({
-    owner: github.context.repo.owner,
-    repo: github.context.repo.repo,
-    ref: ref,
-  });
-
-  const { data: commit } = await client.git.getCommit({
-    owner: github.context.repo.owner,
-    repo: github.context.repo.repo,
+  const { data: refData } = await client.git.getRef({ owner, repo, ref });
+  const { data: commit } = await client.git.getCommit({ owner, repo,
     commit_sha: refData.object.sha,
   });
 
   return commit;
 }
 
-function getCommitUrl(commitSha: string): string {
-  const API_URL_START = "https://api.github.com/repos/";
-  const SITE_URL_START = "https://github.com/";
-  const API_URL_END = "/commits{/sha}";
-  const SITE_URL_END = "/commit";
+async function getContentFromRepo(
+  client: Octokit,
+  path: string,
+): Promise<RepoContents> {
+  const { data } = await client.repos.getContent({ owner, repo, path,
+    ref: github.context.sha,
+  });
 
-  // https://api.github.com/repos/{user}/{repo}/git/commits{/sha}
-  const apiCommitsUrl: string = github.context.payload.repository?.commits_url;
-
-  // https://github.com/{user}/{repo}/git/commits{/sha}
-  const mixedUrl: string = apiCommitsUrl.replace(API_URL_START, SITE_URL_START);
-
-  // https://github.com/{user}/{repo}/git/commit
-  const baseCommitUrl: string = mixedUrl.replace(API_URL_END, SITE_URL_END);
-
-  // https://github.com/{user}/{repo}/git/commit/{commitSha}
-  return `${baseCommitUrl}/${commitSha}`;
+  return data;
 }
 
 
@@ -53,31 +56,24 @@ export async function commitFiles(
 ): Promise<CommitInfo> {
   const previousCommit = await getCommitAt(client, ref);
 
-  const { data: newTree } = await client.git.createTree({
-    owner: github.context.repo.owner,
-    repo: github.context.repo.repo,
+  const { data: newTree } = await client.git.createTree({ owner, repo,
     base_tree: previousCommit.tree.sha,
     tree: blobs,
   });
 
-  const { data: newCommit } = await client.git.createCommit({
-    owner: github.context.repo.owner,
-    repo: github.context.repo.repo,
+  const { data: newCommit } = await client.git.createCommit({ owner, repo,
     message: commitMessage,
     tree: newTree.sha,
     parents: [previousCommit.sha],
   });
 
-  const { data: result } = await client.git.updateRef({
-    owner: github.context.repo.owner,
-    repo: github.context.repo.repo,
-    ref: ref,
+  const { data: result } = await client.git.updateRef({ owner, repo, ref,
     sha: newCommit.sha,
   });
 
   return {
     sha: result.object.sha,
-    url: getCommitUrl(result.object.sha),
+    url: `https://github.com/${owner}/${repo}/git/commit/${result.object.sha}`,
   };
 }
 
@@ -87,9 +83,7 @@ export async function createBlob(
   data: string,
   encoding: string,
 ): Promise<GitBlob> {
-  const { data: fileBlob } = await client.git.createBlob({
-    owner: github.context.repo.owner,
-    repo: github.context.repo.repo,
+  const { data: fileBlob } = await client.git.createBlob({ owner, repo,
     content: data,
     encoding: encoding,
   });
@@ -107,9 +101,7 @@ export async function createComment(
   prNumber: number,
   comment: string,
 ): Promise<void> {
-  await client.issues.createComment({
-    owner: github.context.repo.owner,
-    repo: github.context.repo.repo,
+  await client.issues.createComment({ owner, repo,
     issue_number: prNumber,
     body: comment,
   });
@@ -119,9 +111,7 @@ export async function getCommitFiles(
   client: Octokit,
   sha: string,
 ): Promise<GitFileInfo[]> {
-  const { data } = await client.repos.getCommit({
-    owner: github.context.repo.owner,
-    repo: github.context.repo.repo,
+  const { data } = await client.repos.getCommit({ owner, repo,
     ref: sha,
   });
 
@@ -139,23 +129,47 @@ export async function getCommitMessage(
   return message;
 }
 
+export async function getContent(
+  client: Octokit,
+  path: string,
+): Promise<GitObjectInfo[]> {
+  const data = await getContentFromRepo(client, path) as DirContents;
+  return data.map((item) => ({
+    path: item.path,
+    type: item.type,
+  }));
+}
+
+export async function getDefaultBranch(client: Octokit): Promise<string> {
+  const { data } = await client.repos.get({ owner, repo });
+  return data.default_branch;
+}
+
+export async function getFile(
+  client: Octokit,
+  path: string,
+): Promise<GitFileData> {
+  const fileDetails = await getContentFromRepo(client, path) as FileContents;
+  return {
+    content: fileDetails.content,
+    encoding: fileDetails.encoding,
+    path: fileDetails.path,
+  };
+}
+
 export async function getPrComments(
   client: Octokit,
   prNumber: number,
 ): Promise<string[]> {
   const PER_PAGE = 100;
 
-  const { data } = await client.pulls.get({
-    owner: github.context.repo.owner,
-    repo: github.context.repo.repo,
+  const { data } = await client.pulls.get({ owner, repo,
     pull_number: prNumber,
   });
 
   const prComments: string[] = [];
   for (let i = 0; i < Math.ceil(data.comments / PER_PAGE); i++) {
-    const { data: comments } = await client.issues.listComments({
-      owner: github.context.repo.owner,
-      repo: github.context.repo.repo,
+    const { data: comments } = await client.issues.listComments({ owner, repo,
       issue_number: prNumber,
       per_page: PER_PAGE,
       page: i,
@@ -167,32 +181,11 @@ export async function getPrComments(
   return prComments.reverse();
 }
 
-export async function getPrFile(
-  client: Octokit,
-  path: string,
-): Promise<GitFileData> {
-  const fileContents = await client.repos.getContent({
-    owner: github.context.repo.owner,
-    repo: github.context.repo.repo,
-    path: path,
-    ref: github.context.sha,
-  });
-
-  const fileDetails = fileContents.data[0] || fileContents.data;
-  return {
-    content: fileDetails.content,
-    encoding: fileDetails.encoding,
-    path: fileDetails.path,
-  };
-}
-
 export async function getPrFiles(
   client: Octokit,
   prNumber: number,
 ): Promise<GitFileInfo[]> {
-  const prFilesDetails = await client.pulls.listFiles({
-    owner: github.context.repo.owner,
-    repo: github.context.repo.repo,
+  const prFilesDetails = await client.pulls.listFiles({ owner, repo,
     pull_number: prNumber,
   });
 
@@ -209,11 +202,4 @@ export function getPrNumber(): number {
   }
 
   return pullRequest.number;
-}
-
-export async function getRepoFile(
-  client: Octokit,
-  path: string,
-): Promise<GitFileData> {
-  return await getPrFile(client, path);
 }