+++ test/mocks/github-api.mock.ts
@@ -1,24 +1,34 @@
-import * as core from "./@actions/core.mock";
+import { Octokit } from "@octokit/core";
+
 import * as github from "./@actions/github.mock";
 
-import {
-  COMMIT_MODE_FILE,
-  COMMIT_TYPE_BLOB,
-  INPUT_NAME_REPO_TOKEN,
-} from "../../src/constants";
-import { GitFileData } from "../../src/types";
+import { COMMIT_MODE_FILE, COMMIT_TYPE_BLOB } from "../../src/constants";
+import { GitFileData, GitObjectInfo } from "../../src/types";
 
 
-const token = core.getInput(INPUT_NAME_REPO_TOKEN, { required: true });
-const client = github.getOctokit(token);
+async function _getContent(
+  client: Octokit,
+  path: string,
+): Promise<GitFileData | GitObjectInfo[]> {
+  const { data } = await client.repos.getContent({
+    owner: "mew",
+    repo: "svg-action",
+    path,
+    ref: "heads/ref",
+  });
 
-async function getContent(path: string): Promise<GitFileData> {
-  const { data } = await client.repos.getContent({ path });
-  return {
-    path: data.path,
-    content: data.content,
-    encoding: data.encoding,
-  };
+  if (Array.isArray(data)) {
+    return data.map((item) => ({
+      path: item.path,
+      type: item.type,
+    }));
+  } else {
+    return {
+      path: data.path,
+      content: data.content,
+      encoding: data.encoding,
+    };
+  }
 }
 
 
@@ -42,8 +52,13 @@ export const createComment = jest.fn()
   .mockName("github-api.createComment");
 
 export const getCommitFiles = jest.fn()
-  .mockImplementation(async (_, sha) => {
-    const { data } = await client.repos.getCommit({ ref: sha });
+  .mockImplementation(async (client, sha) => {
+    const { data } = await client.repos.getCommit({
+      owner: "mew",
+      repo: "svg-action",
+      ref: sha,
+    });
+
     return data.files.map((details) => ({
       path: details.filename,
       status: details.status,
@@ -55,17 +70,27 @@ export const getCommitMessage = jest.fn()
   .mockResolvedValue("This is a commit message")
   .mockName("github-api.getCommitMessage");
 
+export const getContent = jest.fn()
+  .mockImplementation(_getContent)
+  .mockName("github-api.getContent");
+
+export const getDefaultBranch = jest.fn()
+  .mockReturnValue("main")
+  .mockName("github-api.getDefaultBranch");
+
+export const getFile = jest.fn()
+  .mockImplementation(_getContent)
+  .mockName("github-api.getFile");
+
 export const getPrComments = jest.fn()
   .mockResolvedValue([])
-  .mockName("github-api.getPrFile");
-
-export const getPrFile = jest.fn()
-  .mockImplementation(async (_, path) => await getContent(path))
-  .mockName("github-api.getPrFile");
+  .mockName("github-api.getPrComment");
 
 export const getPrFiles = jest.fn()
-  .mockImplementation(async (_, prNumber) => {
+  .mockImplementation(async (client, prNumber) => {
     const { data } = await client.pulls.listFiles({
+      owner: "mew",
+      repo: "svg-action",
       pull_number: prNumber,
     });
 
@@ -79,7 +104,3 @@ export const getPrFiles = jest.fn()
 export const getPrNumber = jest.fn()
   .mockReturnValue(github.PR_NUMBER.NO_CHANGES)
   .mockName("github-api.getPrNumber");
-
-export const getRepoFile = jest.fn()
-  .mockImplementation(async (_, path) => await getContent(path))
-  .mockName("github-api.getPrFile");