+++ test/mocks/push.mock.ts
@@ -1,1 +1,0 @@
-export default jest.fn().mockName("commits.ts::main");